refactor module poc to modlifecycle
